Initialize i18n globally as singleton
